rk: ion: rockchip: support sync driection API
If we use ION_FLAG_CACHED_NEEDS_SYNC flag to allocate a
non-coherence ion buffer. User should call ION_IOC_SYNC to
guarantee the coherence of ion buffer.

Before this patch, when user call ION_IOC_SYNC, the ion kernel
driver force cache write-back. It is wrong if a device has written
the ion buffer and the cache should be invalid.

So in this patch, we add a new ioctl ION_IOC_SYNC_DIRECTION for
this case. With this ioctl, user can control the sync driection to
flush or invalidate cache.

Change-Id: Ifc5124f961d4c7541e8935eb8c2202c38b4ec730
Signed-off-by: Zhichao Yu <zhichao.yu@rock-chips.com>